move tests to surveyManager
